Build with -Wall -Wextra, and clear the warnings
The build passed only -Wfatal-errors, so the whole -Wall class was invisible.
Two of the bugs in the last round were exactly what these warnings point at:
min(signed, unsigned) comparing through the usual arithmetic conversions (#74),
and pow() squaring at the argument's width (#73).

Turning them on reports 164 warnings over 20 distinct sites, all fixed here:

- signed/unsigned loop and comparison mismatches: index vectors with size_t in
  expr.h and integer_backend.h; hold num_bits() in an int64_t in pow_mod() and
  pow(rational, integer); compare `rounds` against primes.size() as a size_t,
  and reject a negative round count while there;
- real<10>'s formatter compares against -a.exp, which is now taken as a size_t
  once, with the digit position derived by subtraction instead of by adding a
  negative;
- unused parameters: real's normalize-skipping constructor and the builtin
  stre() overload lose their parameter names;
- operator"" _u128 in the tests loses the deprecated space;
- popcount and modN test assertions compare like with like instead of relying on
  a conversion;
- TEST(0 <= am) in the long stress test asserted nothing, since am is unsigned.

No -Werror: a new compiler version's new warnings should not break the build.

REVIEW2.md section 8.1.

+++ algebra/__stress_test.h
@@ -93,7 +93,7 @@ void catch_exceptions(const M<T>& m, const auto& fn) {
     }
 }
 
-std::string shorten(const std::string& a, int pre, int post) {
+std::string shorten(const std::string& a, size_t pre, size_t post) {
     return (a.size() <= pre + post) ? a : (a.substr(0, pre) + "..." + a.substr(a.size() - post));
 }
 
@@ -214,7 +214,7 @@ void stress_test(const auto& fn, int max_errors = 6) {
     };
 
     std::vector<std::thread> threads;
-    for (int i = 0; i < std::thread::hardware_concurrency(); i++)
+    for (unsigned i = 0; i < std::thread::hardware_concurrency(); i++)
         threads.push_back(std::thread(func));
     threads[0].join();
 }
